move endpoint stdlib to functions pkg
