+++ runtime/protocol/http/common/test/aws/smithy/kotlin/runtime/http/endpoints/functions/FunctionsTest.kt
@@ -2,11 +2,11 @@
  * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
  * SPDX-License-Identifier: Apache-2.0.
  */
-package aws.smithy.kotlin.runtime.http.endpoints.util
+package aws.smithy.kotlin.runtime.http.endpoints.functions
 
 import kotlin.test.*
 
-class UtilTest {
+class FunctionsTest {
     @Test
     fun testSubstringInvalidIndices() =
         assertNull(substring("foo", 2, 1, false))
